Taking out radio buttons. Implementing selection and filtering logic
I removed the radio buttons because the selection feature would be
simpler. I then created the filtering logic that will allow the user to
select a departure and arrival location and then view the flights that
fit the locations specified. Addiitonally, I made it so when you select
a table entry and then click the select flight button, it will return
the flight number to the controller so that it can give that to the seat
controller.